# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
        "/applications": {
            "get": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "List job applications",
                "description": "This route list all registered job applications.",
                "operationId": "dbd30e586e706d8c52bd39521719bd15",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "resume_id": 1,
                                        "resume_candidate": "Cicero Augusto",
                                        "resume_description": "System Analyst with experience"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/applications/job/{id}": {
            "get": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Search job application by job id",
                "description": "This route gets the data of a job application through a job id.",
                "operationId": "5db22e31fda5540f06a829f0ad9fc3e5",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "resume_id": 1,
                                        "resume_candidate": "Cicero Augusto",
                                        "resume_description": "System Analyst with experience"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/applications/candidate/{name}": {
            "get": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Search job application by candidate name",
                "description": "This route gets the data of a job application through a candidate name.",
                "operationId": "768ae14c3d40cc9ad2f9916178dfc1b4",
                "parameters": [
                    {
                        "name": "Name",
                        "in": "path",
                        "description": "JobApplication Candidate Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "resume_id": 1,
                                        "resume_candidate": "Cicero Augusto",
                                        "resume_description": "System Analyst with experience"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "List jobs",
                "description": "List all resgistered jobs.",
                "operationId": "61095dd3fdb7b25e7daa0084b396ca6e",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        {
                                            "job_id": 379,
                                            "job_title": "Analyst",
                                            "job_description": "System Analyst with experience",
                                            "created_at": "2022-02-18 12:13:59",
                                            "job_location": "Brazil",
                                            "job_expires": "2022-03-20",
                                            "application_email": "hr@bigcompany.com",
                                            "company_name": "Big Company",
                                            "company_website": "http://www.bigcompany.com.br",
                                            "company_tagline": "",
                                            "company_twitter": "@bigcompany",
                                            "company_video": "",
                                            "application_deadline": "2022-02-03",
                                            "company_logo": "http://localhost/wp-content/uploads/2022/02/bigcompany.png",
                                            "extra_fields": {
                                                "experience_needed": "Yes"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/jobs/{ID}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Search job by ID",
                "description": "Gets the data of a job through an ID.",
                "operationId": "59cfd805e4e8cbae6213c2cdcc5d4d33",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "job_id": 1,
                                        "job_title": "Analyst",
                                        "job_description": "System Analyst with experience",
                                        "created_at": "2022-02-18 12:13:59",
                                        "job_location": "Brazil",
                                        "job_expires": "2022-03-20",
                                        "application_email": "hr@bigcompany.com",
                                        "company_name": "Big Company",
                                        "company_website": "http://www.bigcompany.com.br",
                                        "company_tagline": "",
                                        "company_twitter": "@bigcompany",
                                        "company_video": "",
                                        "application_deadline": "2022-02-03",
                                        "company_logo": "http://localhost/wp-content/uploads/2022/02/bigcompany.png",
                                        "extra_fields": {
                                            "experience_needed": "Yes"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}